+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
  - `createClient()` uses default Memcached address
  - don't allow new requests when *is ending* or *is closed*
  - write response resolving
+ - wrong commands are rejected immediately
  
 ### Added:
  - client can emit events (`close`)